request body should be compressed earlier for expect 100-continue (to update the content length in header)
