+++ Titanium.Web.Proxy/Helpers/HttpWriter.cs
@@ -94,12 +94,9 @@ namespace Titanium.Web.Proxy.Helpers
         /// <returns></returns>
         public async Task WriteHeadersAsync(HeaderCollection headers, bool flush = true)
         {
-            if (headers != null)
+            foreach (var header in headers)
             {
-                foreach (var header in headers)
-                {
-                    await header.WriteToStreamAsync(this);
-                }
+                await header.WriteToStreamAsync(this);
             }
 
             await WriteLineAsync();
@@ -274,52 +271,12 @@ namespace Titanium.Web.Proxy.Helpers
         /// <returns></returns>
         protected async Task WriteAsync(RequestResponseBase requestResponse, bool flush = true)
         {
-            if (requestResponse.HasBody)
+            var body = requestResponse.CompressBodyAndUpdateContentLength();
+            await WriteHeadersAsync(requestResponse.Headers, flush);
+
+            if (body != null)
             {
-                bool isChunked = requestResponse.IsChunked;
-                string contentEncoding = requestResponse.ContentEncoding;
-
-                var body = requestResponse.Body;
-                if (contentEncoding != null && body != null)
-                {
-                    body = GetCompressedBody(contentEncoding, body);
-
-                    if (isChunked == false)
-                    {
-                        requestResponse.ContentLength = body.Length;
-                    }
-                    else
-                    {
-                        requestResponse.ContentLength = -1;
-                    }
-                }
-
-                await WriteHeadersAsync(requestResponse.Headers, flush);
-                await WriteBodyAsync(body, isChunked);
-            }
-            else
-            {
-                await WriteHeadersAsync(requestResponse.Headers, flush);
-            }
-        }
-
-        /// <summary>
-        /// get the compressed body from given bytes
-        /// </summary>
-        /// <param name="encodingType"></param>
-        /// <param name="body"></param>
-        /// <returns></returns>
-        internal byte[] GetCompressedBody(string encodingType, byte[] body)
-        {
-            var compressor = CompressionFactory.GetCompression(encodingType);
-            using (var ms = new MemoryStream())
-            {
-                using (var zip = compressor.GetStream(ms))
-                {
-                    zip.Write(body, 0, body.Length);
-                }
-
-                return ms.ToArray();
+                await WriteBodyAsync(body, requestResponse.IsChunked);
             }
         }
     }
