+++ Titanium.Web.Proxy/Http/RequestResponseBase.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.IO;
 using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
+using Titanium.Web.Proxy.Compression;
 using Titanium.Web.Proxy.Extensions;
 using Titanium.Web.Proxy.Helpers;
 using Titanium.Web.Proxy.Models;
@@ -149,6 +150,55 @@ namespace Titanium.Web.Proxy.Http
         internal bool OriginalHasBody;
 
         internal abstract void EnsureBodyAvailable(bool throwWhenNotReadYet = true);
+
+        /// <summary>
+        /// get the compressed body from given bytes
+        /// </summary>
+        /// <param name="encodingType"></param>
+        /// <param name="body"></param>
+        /// <returns></returns>
+        internal byte[] GetCompressedBody(string encodingType, byte[] body)
+        {
+            var compressor = CompressionFactory.GetCompression(encodingType);
+            using (var ms = new MemoryStream())
+            {
+                using (var zip = compressor.GetStream(ms))
+                {
+                    zip.Write(body, 0, body.Length);
+                }
+
+                return ms.ToArray();
+            }
+        }
+
+        internal byte[] CompressBodyAndUpdateContentLength()
+        {
+            bool isChunked = IsChunked;
+            string contentEncoding = ContentEncoding;
+
+            if (HasBody)
+            {
+                var body = Body;
+                if (contentEncoding != null && body != null)
+                {
+                    body = GetCompressedBody(contentEncoding, body);
+
+                    if (isChunked == false)
+                    {
+                        ContentLength = body.Length;
+                    }
+                    else
+                    {
+                        ContentLength = -1;
+                    }
+                }
+
+                return body;
+            }
+
+            ContentLength = 0;
+            return null;
+        }
 
         /// <summary>
         /// Body as string
